Replace `vim-caser` with `vim-abolish`
See https://github.com/tpope/vim-abolish.

From the docs:

> Want to turn `fooBar` into `foo_bar`? Press `crs` (coerce to
snake_case). MixedCase (`crm`), camelCase (`crc`), UPPER_CASE (`cru`),
dash-case (`cr-`), and dot.case (`cr.`) are all just 3 keystrokes away.

The substitution functionality looks interesting, too:

> From a conceptual level, one way to think about how this substitution
works is to imagine that in the braces you are declaring the
requirements for turning that word from singular to plural. In the
facility example, the same base letters in both the singular and plural
form of the word are `facilit` To turn "facility" to a plural word you
must change the `y` to `ies` so you specify `{y,ies}` in the braces.

> To convert the word "building" from singular to plural, again look at
the common letters between the singular and plural forms: `building.` In
this case you do not need to remove any letter from building to turn it
into plural form and you need to add an `s` so the braces should be
`{,s}.`