Re-add vim-projectionist
Whilst I wasn't able to get it to work for all my use cases previously,
it is useful for creating alternative files in Fractal, such as
switching between the `*.config.yml` and `*.twig` files for a component.

Other use cases, such as creating alternative files for classes and
tests in PHP and Drupal projects, I can look at again now the plugin is
installed.