Add `vim-autoread`
> Automatically causes vim to reload files which have been written on
disk but not modified in the buffer since the last write from vim. This
enables a file open in vim to be edited using another application and
saved. Upon returning to vim, as long as you haven't modified the file
since the last change, the file will be automatically updated to reflect
the changes made on disk, as though you had pressed :e manually.

https://github.com/djoshea/vim-autoread