Add edit_alternate.vim
https://github.com/tjdevries/edit_alternate.vim

This works for switching between classes and tests in Drupal modules,
which is something I was struggling to do with Projectionist.

Different to Projectionist, though, this plugin doesn't create an
alternate file if it doesn't exist, and doesn't prompt to select from
multiple options and returns the first matching one.